Added C header file containing the DASSL function prototype.
This is needed by Cython to expose DASSL to Python. The ddassl_() function
prototype is consistent with that expected by Fortran, i.e. the parameters
are passed by reference rather than by value, so they must all be pointers.

Note that this header file is safe to use in C++ as well, though we will not
need that for our purposes.