Added Cython module for wrapping DASSL code and setup.py build script.
The Cython module is not written with a pure Python mode because it must be
compiled to make the DASSL functionality available.

Again, depending on your platform you may need to tweak the setup.py file a
bit. Of particular note is the libraries to link with when mixing compiled
Fortran and C code.