Added unit test module and first unit test for DASSL.
All unit tests should be placed in test.py. As of now there is only one unit
test, which is copied from the rxnSeries example. Instead of producing a plot,
however, the unit test compares the numerical solution to the known analytical
solution.

As of this commit the unit test passes successfully.