Reorganizing the documentation to match other projects of mine.
The previous documentation was very informal, and was created back when I
thought the number of users of this code would be very small. Since I am now
getting interest from users that don't sit in my office, it is time to
formalize the documentation, both for their sake and for mine!

The new documentation will have separate documents for end users and
developers.