New setup.py script and Makefile for building PyDAS using f2py.
The build scripts are much simpler than before because we are relying on
distutils to compile DASSL, DASPK, and DASKR as well as their Python
wrappers. This lets distutils worry about choosing the appropriate Fortran
and C compilers, and thereby hopefully making PyDAS easier to install,
especially with non-GCC and non-MinGW compilers.