Updated .gitignore to ignore new temporary build files.
Now that the build process has changed from Cython to f2py, the temporary
build files generated have changed a bit. By changing to f2py, more of the
intermediate build files are made in the build folder, but a few of the
f2py-generated wrapper files are created next to the signature files
(presumably because they are *.f and *.c "source" files). This commit
updates .gitignore accordingly.