Tweak of DASSL signature to enable residual and Jacobian to work properly.
The problem is that the residual and jacobian external subroutines do not
have the number of equations as a parameter, which makes it difficult for
f2py to wrap them properly. To get around this, we pass this value as the
first item in the ipar array; since we aren't using rpar and ipar, this is
an okay thing to do.

Note that I got the idea for this from another attempt to wrap DASSL with
f2py:

http://msdl.cs.mcgill.ca/people/slacoste/research/report/SummerReport.html

Hopefully this is the last time anyone has to figure this out!