Moved content of pydas.pyx to pydas.dassl (pure Python) module.
By automatically importing the contents of this module in the pydas
__init__.py file, we won't break the API. Along the way, I've removed all
of the Cython syntax to make this file into a pure Python file. This means
that Cython is no longer a dependency for PyDAS, leaving the end user free
to choose the method of speeding up the residual and/or Jacobian. (Cython
is still highly recommended of course!)