Added wrapper __residual and __jacobian functions to DASSL.
These appear to be necessary in order for derived classes to be able to
write their residual and jacobian functions in Cython. Since Cython is
currently our preferred means of speeding up Python code, we still want to
be able to use it even with the f2py-wrapped Fortran code. My only concern
is the speed penalty of having to pass through a Python layer for every
call to the residual or jacobian functions.