Add the CALL command to batch file calls.
Anaconda Python provides mingw, but the executables (gcc, gfortran, ar)
are wrapped in batch files, i.e. "gfortran.bat". Because of this, the
first command `%f77% %cflags% -c dassl/daux.f -o dassl/daux.o` will
complete successfully but the next command will not be run. In order
to execute a batch file from within a batch file and *resume*
execution in the original batch file, the CALL command must be used.
Note that in this patch, `CALL python` is not necessary in Anaconda
because by default the python executable is added to the system path.
However, adding it here does no harm, and `make.bat` will continue
to work on systems where the python executable has been wrapped in a
batch file.